feat: split traffic_light result
